# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST multi-session para WhatsApp usando whatsmeow. Baseada na WuzAPI com campos JSON em PascalCase.",
        "title": "FioZap API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "FioZap Support",
            "url": "https://github.com/fiozap/fiozap"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/sessions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Lista todas as sessoes WhatsApp criadas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Listar sessoes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.SessionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria uma nova sessao WhatsApp com o nome informado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Criar sessao",
                "parameters": [
                    {
                        "description": "Nome da sessao",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna informacoes de uma sessao especifica",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Obter sessao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove sessao e todos os dados associados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Deletar sessao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ActionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/sessions/{name}/connect": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Inicia conexao da sessao com WhatsApp (gera QR code se necessario)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Conectar sessao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/sessions/{name}/disconnect": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Desconecta a sessao do WhatsApp (mantem dados)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Desconectar sessao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/sessions/{name}/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Faz logout da sessao (requer novo QR code para reconectar)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Logout da sessao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ActionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/sessions/{name}/qr": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna QR code para autenticacao. Use ?format=image para PNG",
                "produces": [
                    "application/json",
                    "image/png"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Obter QR code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "image"
                        ],
                        "type": "string",
                        "description": "Formato: image para PNG",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.QRResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ActionResponse": {
            "type": "object",
            "properties": {
                "Details": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "dto.CreateSessionRequest": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string"
                },
                "Provider": {
                    "description": "\"whatsmeow\" (default) ou \"cloudapi\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ProviderType"
                        }
                    ]
                }
            }
        },
...
        "dto.ProviderType": {
            "type": "string",
            "enum": [
                "whatsmeow",
                "cloudapi"
            ],
            "x-enum-comments": {
                "ProviderCloudAPI": "API oficial Meta",
                "ProviderWhatsmeow": "default - nao-oficial"
            },
            "x-enum-descriptions": [
                "default - nao-oficial",
                "API oficial Meta"
            ],
            "x-enum-varnames": [
                "ProviderWhatsmeow",
                "ProviderCloudAPI"
            ]
        },
        "dto.QRResponse": {
            "type": "object",
            "properties": {
                "QRCode": {
                    "type": "string"
                }
            }
        },
...
        "dto.Response": {
            "description": "Resposta padrao da API FioZap",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "string",
                    "example": "error message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.SessionResponse": {
            "type": "object",
            "properties": {
                "Connected": {
                    "type": "boolean"
                },
                "JID": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                },
                "Phone": {
                    "type": "string"
                },
                "Provider": {
                    "$ref": "#/definitions/dto.ProviderType"
                },
                "PushName": {
                    "type": "string"
                },
                "Token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token global (GLOBAL_API_TOKEN) ou token da sessao",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}